# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Spatial Memory Network - an AR app that lets users pin multimedia memories to real-world geographic locations.",
        "title": "Spatial Memory Network API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/spatial-memory/spatial-memory/issues",
            "email": "support@spatial-memory.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/moderation/queue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the list of memories pending moderation review",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List moderation queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status: pending, approved, rejected",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moderation queue",
                        "schema": {
                            "$ref": "#/definitions/response.PaginatedResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/moderation/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get statistics about the moderation queue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get moderation statistics",
                "responses": {
                    "200": {
                        "description": "Moderation statistics",
                        "schema": {
                            "$ref": "#/definitions/model.ModerationStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/moderation/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single moderation item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get moderation item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Moderation item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moderation item",
                        "schema": {
                            "$ref": "#/definitions/model.ModerationItem"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/moderation/{id}/review": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually approve or reject a moderation item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manual review",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Moderation item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review decision",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ManualReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review processed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ManualReviewRequest": {
            "type": "object",
            "required": [
                "approved"
            ],
            "properties": {
                "approved": {
                    "type": "boolean"
                },
                "note": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "model.ModerationItem": {
            "type": "object",
            "properties": {
                "ai_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ai_safe_score": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "memory_content": {
                    "type": "string"
                },
                "memory_id": {
                    "type": "integer"
                },
                "memory_title": {
                    "description": "Joined fields",
                    "type": "string"
                },
                "report_count": {
                    "type": "integer"
                },
                "review_note": {
                    "type": "string"
                },
                "reviewed_at": {
                    "type": "string"
                },
                "reviewer_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.ModerationStatus"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_nickname": {
                    "type": "string"
                }
            }
        },
        "model.ModerationStats": {
            "type": "object",
            "properties": {
                "approved_count": {
                    "type": "integer"
                },
                "escalated_count": {
                    "type": "integer"
                },
                "pending_count": {
                    "type": "integer"
                },
                "rejected_count": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "model.ModerationStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "ModerationStatusPending",
                "ModerationStatusApproved",
                "ModerationStatusRejected",
                "ModerationStatusEscalated"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 40000
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "response.PaginatedData": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/response.PaginatedData"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}